Port ST_CrossesDateLine from Predicates to Functions
