+++ flink/src/main/java/org/apache/sedona/flink/Catalog.java
@@ -53,6 +53,7 @@ public class Catalog {
                 new Functions.ST_CollectionExtract(),
                 new Functions.ST_ConcaveHull(),
                 new Functions.ST_ConvexHull(),
+                new Functions.ST_CrossesDateLine(),
                 new Functions.ST_Envelope(),
                 new Functions.ST_Difference(),
                 new Functions.ST_Dimension(),
@@ -155,7 +156,6 @@ public class Catalog {
                 new Predicates.ST_Intersects(),
                 new Predicates.ST_Contains(),
                 new Predicates.ST_Crosses(),
-                new Predicates.ST_CrossesDateLine(),
                 new Predicates.ST_Within(),
                 new Predicates.ST_Covers(),
                 new Predicates.ST_CoveredBy(),
