+++ flink/src/main/java/org/apache/sedona/flink/expressions/Functions.java
@@ -164,6 +164,21 @@ public class Functions {
         }
     }
 
+    public static class ST_CrossesDateLine extends ScalarFunction
+    {
+        /**
+         * Constructor for relation checking without duplicate removal
+         */
+        public ST_CrossesDateLine() {}
+
+        @DataTypeHint("Boolean")
+        public Boolean eval(@DataTypeHint(value = "RAW", bridgedTo = org.locationtech.jts.geom.Geometry.class) Object o)
+        {
+            Geometry geom = (Geometry) o;
+            return org.apache.sedona.common.Functions.crossesDateLine(geom);
+        }
+    }
+
     public static class ST_Envelope extends ScalarFunction {
         @DataTypeHint(value = "RAW", bridgedTo = org.locationtech.jts.geom.Geometry.class)
         public Geometry eval(@DataTypeHint(value = "RAW", bridgedTo = org.locationtech.jts.geom.Geometry.class) Object o) {
