+++ python/tests/sql/test_dataframe_api.py
@@ -86,6 +86,7 @@ test_configurations = [
     (stf.ST_ConcaveHull, ("geom", 1.0, True), "triangle_geom", "", "POLYGON ((1 1, 1 0, 0 0, 1 1))"),
     (stf.ST_ConvexHull, ("geom",), "triangle_geom", "", "POLYGON ((0 0, 1 1, 1 0, 0 0))"),
     (stf.ST_CoordDim, ("point",), "point_geom", "", 2),
+    (stp.ST_CrossesDateLine, ("line",), "line_crossing_dateline", "", True),
     (stf.ST_Difference, ("a", "b"), "overlapping_polys", "", "POLYGON ((1 0, 0 0, 0 1, 1 1, 1 0))"),
     (stf.ST_Dimension, ("geom",), "geometry_geom_collection", "", 1),
     (stf.ST_Distance, ("a", "b"), "two_points", "", 3.0),
@@ -168,7 +169,6 @@ test_configurations = [
     # predicates
     (stp.ST_Contains, ("geom", lambda: f.expr("ST_Point(0.5, 0.25)")), "triangle_geom", "", True),
     (stp.ST_Crosses, ("line", "poly"), "line_crossing_poly", "", True),
-    (stp.ST_CrossesDateLine, ("line",), "line_crossing_dateline", "", True),
     (stp.ST_Disjoint, ("a", "b"), "two_points", "", True),
     (stp.ST_Equals, ("line", lambda: f.expr("ST_Reverse(line)")), "linestring_geom", "", True),
     (stp.ST_Intersects, ("a", "b"), "overlapping_polys", "", True),
@@ -241,6 +241,7 @@ wrong_type_configurations = [
     (stf.ST_CollectionExtract, (None,)),
     (stf.ST_ConcaveHull, (None, 1.0)),
     (stf.ST_ConvexHull, (None,)),
+    (stp.ST_CrossesDateLine, (None,)),
     (stf.ST_Difference, (None, "b")),
     (stf.ST_Difference, ("", None)),
     (stf.ST_Distance, (None, "")),
@@ -333,7 +334,6 @@ wrong_type_configurations = [
     (stp.ST_Contains, ("", None)),
     (stp.ST_Crosses, (None, "")),
     (stp.ST_Crosses, ("", None)),
-    (stp.ST_CrossesDateLine, (None,)),
     (stp.ST_Disjoint, (None, "")),
     (stp.ST_Disjoint, ("", None)),
     (stp.ST_Equals, (None, "")),
