+++ spark/common/src/main/scala/org/apache/spark/sql/sedona_sql/expressions/Predicates.scala
@@ -166,14 +166,6 @@ case class ST_Crosses(inputExpressions: Seq[Expression])
   }
 }
 
-case class ST_CrossesDateLine(inputExpressions: Seq[Expression])
-  extends InferredExpression(Predicates.crossesDateLine _) {
-
-  protected def withNewChildrenInternal(newChildren: IndexedSeq[Expression]) = {
-    copy(inputExpressions = newChildren)
-  }
-}
-
 /**
   * Test if leftGeometry overlaps rightGeometry
   *
